requirements.txt
flask
torch
torchvision
pillow
opencv-python